requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

pydantic>=2.8.0
pydantic-settings>=2.4.0
python-dotenv>=1.0.1

PyMuPDF>=1.24.9
sentence-transformers>=3.0.1
numpy>=1.26.4
faiss-cpu>=1.8.0.post1

openai>=1.40.0

pytest>=8.3.0
httpx>=0.27.0
